Neatened some things up
