+++ app/components/ToggleButton.jsx
@@ -1,13 +1,13 @@
 import styles from "./styles/ToggleButton.module.css";
 
-export default function ToggleButton({ states, value, onChange, className }) {
+export default function ToggleButton({ states, currentState, onChange, className }) {
 
 	return (
 		<button className={`${styles.toggleButton} ${className}`} onClick={() => {
-			const nextState = (value + 1) % states.length;
+			const nextState = (currentState + 1) % states.length;
 			onChange(nextState);
 		}}>
-			{states[value]}
+			{states[currentState]}
 		</button>
 	);
 }
