+++ app/components/styles/SongTitle.module.css
@@ -9,6 +9,6 @@
 
 @media (max-width: 425px) {
 	.title {
-		font-size: 0.7rem;
+		font-size: 0.65rem;
 	}
 }
