+++ app/music/[id]/page.module.css
@@ -1,47 +1,63 @@
-.container {
- 	backdrop-filter: blur(80px);
-	z-index: 2;
-}
-
 .main {
-	position: absolute;
-	top: 50%;
-	left: 50%;
-	translate: -50% -50%;
-	z-index: 3;
+	position: relative;
+	z-index: 2;
 	display: grid;
 	grid-template-columns: auto 1fr 1fr;
-	grid-template-rows: 1fr 1fr;
-	gap: 15px;
-	width: 90%;
-	height: 80%;
-	margin-top: 80px;
+	grid-template-rows: auto 1fr;
+	gap: 20px;
+	padding: 100px;
+	/* width: 100%; */
+	overflow: scroll;
 }
 
 .songCover {
+	grid-row: 1;
+	grid-column: 1;
 	width: 250px;
 	height: 250px;
 	border: 1px solid var(--border-color);
 }
 
 .title {
- margin-top: 10px;
- text-wrap: nowrap;
+	grid-column: 2;
+	margin-left: 10px;
+	text-wrap: nowrap;
+}
+
+.songControls {
+	justify-self: center;
+	position: absolute;
+	bottom: 25px;
+	gap: 10px;
+	width: 80%
 }
 
 @media (max-width: 425px) { 
 	.main {
-		grid-template-columns: 1fr;
-		grid-template-rows: 1fr 1fr;
+		padding: 0;
+		padding-top: 100px;
+		grid-template-columns: 100%;
+		grid-template-rows: 250px 50px 50px;
 	}
 
 	.main :global(.SongCover) {
 		margin: auto;
 	}
 
+	.songControls {
+		width: 80%; 
+		margin-top: 100px;
+	}
+
+	.songCover {
+		grid-row: 1;
+		grid-column: 1;
+	}
+
 	.title {
+		grid-row: 2;
+		grid-column: 1;
 		font-size: 1.5em;
 		text-align: center;
-		margin-top: 20px;
 	}
 }
